# docs/api/openapi.yaml
openapi: 3.0.3
info:
    title: Adblock Compiler API
    description: |
        **Compiler-as-a-Service** for adblock filter lists. Transform, optimize, and combine filter lists from multiple sources with real-time progress tracking.

        ## Features
        - 🎯 Multi-Source Compilation
        - ⚡ Performance (Gzip compression, caching, request deduplication)
        - 🔄 Circuit Breaker with retry logic
        - 📊 Visual Diff between compilations
        - 📡 Real-time progress via SSE and WebSocket
        - 🎪 Batch Processing
        - 🌍 Universal (Deno, Node.js, Cloudflare Workers, browsers)

        ## Links
        - [GitHub Repository](https://github.com/jaypatrick/adblock-compiler)
        - [Documentation](https://github.com/jaypatrick/adblock-compiler/tree/master/docs)
        - [Web UI](https://adblock-compiler.jayson-knight.workers.dev/)

    version: 2.0.0
    license:
        name: GPL-3.0
        url: https://github.com/jaypatrick/adblock-compiler/blob/master/LICENSE
    contact:
        name: Jayson Knight
        url: https://github.com/jaypatrick/adblock-compiler
servers:
    - url: https://adblock-compiler.jayson-knight.workers.dev
      description: Production server
    - url: http://localhost:8787
      description: Local development server
...
paths:
...
    /compile:
        post:
            tags: [Compilation]
            summary: Compile filter list (JSON)
            description: |
                Compile filter lists and return results as JSON. Results are cached for 1 hour.
                Supports request deduplication for concurrent identical requests.
            operationId: compileJson
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CompileRequest'
                        examples:
                            simple:
                                summary: Simple compilation
                                value:
                                    configuration:
                                        name: My Filter List
                                        sources:
                                            - source: https://example.com/filters.txt
                                              transformations:
                                                  - RemoveComments
                                                  - Validate
                                        transformations:
                                            - Deduplicate
                            withPreFetched:
                                summary: With pre-fetched content
                                value:
                                    configuration:
                                        name: Custom Rules
                                        sources:
                                            - source: my-custom-rules
                                    preFetchedContent:
                                        my-custom-rules: "||ads.example.com^\n||tracking.example.com^"
            responses:
                '200':
                    description: Compilation successful
                    headers:
                        X-Cache:
                            schema:
                                type: string
                                enum: [HIT, MISS]
                            description: Cache hit/miss status
                        X-Request-Deduplication:
                            schema:
                                type: string
                                enum: [HIT]
                            description: Present if request was deduplicated
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CompileResponse'
                '400':
                    $ref: '#/components/responses/BadRequestError'
                '429':
                    $ref: '#/components/responses/RateLimitError'
                '500':
                    $ref: '#/components/responses/CompilationError'
    /compile/stream:
        post:
            tags: [Streaming]
            summary: Compile with real-time progress (SSE)
            description: |
                Compile filter lists with real-time progress updates via Server-Sent Events.
                Streams events including source downloads, transformations, diagnostics, cache operations, network events, and metrics.
            operationId: compileStream
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CompileRequest'
            responses:
                '200':
                    description: Event stream
                    content:
                        text/event-stream:
                            schema:
                                type: string
                                description: |
                                    Server-Sent Events stream with the following event types:

                                    **Compilation Events:**
                                    - `log` - Log messages (info, warn, error, debug)
                                    - `source:start` - Source download started
                                    - `source:complete` - Source download completed
                                    - `source:error` - Source download failed
                                    - `transformation:start` - Transformation started
                                    - `transformation:complete` - Transformation completed
                                    - `progress` - Compilation progress update
                                    - `result` - Final compilation result
                                    - `done` - Compilation finished
                                    - `error` - Compilation error

                                    **Enhanced Events:**
                                    - `diagnostic` - Diagnostic event from tracing system
                                    - `cache` - Cache operation (hit/miss/write)
                                    - `network` - Network operation (HTTP requests)
                                    - `metric` - Performance metric
                            examples:
                                logEvent:
                                    value: |
                                        event: log
                                        data: {"level":"info","message":"Fetching source..."}
                                sourceStart:
                                    value: |
                                        event: source:start
                                        data: {"source":{"name":"Source 1","source":"https://..."},"sourceIndex":0,"totalSources":2}
                                diagnostic:
                                    value: |
                                        event: diagnostic
                                        data: {"eventId":"evt-123","timestamp":"2026-01-14T05:00:00Z","category":"compilation","severity":"info","message":"Started compilation"}
                                cache:
                                    value: |
                                        event: cache
                                        data: {"operation":"hit","key":"cache:abc123","size":1024}
                                network:
                                    value: |
                                        event: network
                                        data: {"method":"GET","url":"https://example.com/filters.txt","statusCode":200,"durationMs":234,"responseSize":51200}
                                metric:
                                    value: |
                                        event: metric
                                        data: {"metric":"download_speed","value":218.5,"unit":"KB/s"}
                '400':
                    $ref: '#/components/responses/BadRequestError'
                '429':
                    $ref: '#/components/responses/RateLimitError'
    /compile/batch:
        post:
            tags: [Compilation]
            summary: Batch compile multiple lists
            description: Compile multiple filter lists in parallel (max 10 per batch)
            operationId: compileBatch
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BatchCompileRequest'
            responses:
                '200':
                    description: Batch compilation results
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchCompileResponse'
                '400':
                    $ref: '#/components/responses/BadRequestError'
                '429':
                    $ref: '#/components/responses/RateLimitError'
    /compile/async:
        post:
            tags: [Queue]
            summary: Queue async compilation job
            description: |
                Queue a compilation job for asynchronous processing. Returns immediately with a request ID.
                Use GET /queue/results/{requestId} to retrieve results when complete.
            operationId: compileAsync
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CompileRequest'
            responses:
                '202':
                    description: Job queued successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/QueueResponse'
                '400':
                    $ref: '#/components/responses/BadRequestError'
                '500':
                    $ref: '#/components/responses/ServiceUnavailable'
    /compile/batch/async:
        post:
            tags: [Queue]
            summary: Queue batch async compilation
            description: Queue multiple compilations for async processing
            operationId: compileBatchAsync
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BatchCompileRequest'
            responses:
                '202':
                    description: Batch queued successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/QueueResponse'
                '400':
                    $ref: '#/components/responses/BadRequestError'
                '500':
                    $ref: '#/components/responses/ServiceUnavailable'
...
    /ast/parse:
        post:
            tags: [Compilation]
            summary: Parse filter rules into AST
            description: Parses adblock filter rules into an Abstract Syntax Tree representation
            operationId: parseAST
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ASTParseRequest'
            responses:
                '200':
                    description: Parsed AST
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ASTParseResponse'
                '400':
                    description: Invalid request (must provide rules or text)
                '500':
                    description: Parse error
    /admin/storage/stats:
        get:
            tags: [Admin]
            summary: Get storage statistics
            description: Returns database table counts and expired entry counts
            operationId: getAdminStorageStats
            security:
                - AdminKey: []
            responses:
                '200':
                    description: Storage statistics
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminStorageStatsResponse'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
    /admin/storage/clear-expired:
        post:
            tags: [Admin]
            summary: Clear expired entries
            description: Deletes all expired entries from storage tables
            operationId: clearExpiredEntries
            security:
                - AdminKey: []
            responses:
                '200':
                    description: Entries cleared
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminOperationResponse'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
    /admin/storage/clear-cache:
        post:
            tags: [Admin]
            summary: Clear cache entries
            description: Deletes all cache entries from storage tables
            operationId: clearCacheEntries
            security:
                - AdminKey: []
            responses:
                '200':
                    description: Cache cleared
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminOperationResponse'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
    /admin/storage/export:
        get:
            tags: [Admin]
            summary: Export storage data
            description: Exports storage entries as JSON (includes storage entries, filter cache, and compilation metadata; returned amounts subject to implementation limits)
            operationId: exportStorage
            security:
                - AdminKey: []
            responses:
                '200':
                    description: Storage export
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminExportResponse'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
    /admin/storage/vacuum:
        post:
            tags: [Admin]
            summary: Vacuum database
            description: Runs VACUUM on the D1 database to reclaim space
            operationId: vacuumDatabase
            security:
                - AdminKey: []
            responses:
                '200':
                    description: Vacuum completed
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminOperationResponse'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
    /admin/storage/tables:
        get:
            tags: [Admin]
            summary: List database tables
            description: Returns all tables and indexes in the D1 database
            operationId: listStorageTables
            security:
                - AdminKey: []
            responses:
                '200':
                    description: Table listing
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminTablesResponse'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '503':
                    $ref: '#/components/responses/ServiceUnavailable'
...
    /admin/backends:
        get:
            tags: [Admin]
            summary: Backend health status
            description: Returns health and connectivity status of both D1 and PostgreSQL backends
            operationId: getBackendStatus
            security:
                - AdminKey: []
            responses:
                '200':
                    description: Backend status
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BackendStatusResponse'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
...
components:
...
    securitySchemes:
        Turnstile:
            type: apiKey
            in: header
            name: turnstileToken
            description: Cloudflare Turnstile token (if enabled on the server)
        AdminKey:
            type: apiKey
            in: header
            name: X-Admin-Key
            description: Admin API key for protected admin endpoints
security: []